Extract out mq-cursor and mq-node
They're used in both better-navigation and intellisense
