+++ src/plugins/intellisense/latex-parsing.tsx
@@ -1,6 +1,8 @@
 import { ExpressionAug } from "../../../text-mode-core";
 import { latexStringToIdentifierString } from "./view";
-import { MathQuillField, selectionDirection } from "#components";
+import { MathQuillField } from "#components";
+import { MqNodeViaDom } from "../../mathquill/mq-node";
+import { getCursorHead, MqCursorViaDom } from "../../mathquill/mq-cursor";
 
 export function mapAugAST(
   node: ExpressionAug["latex"],
@@ -26,137 +28,6 @@ export function mapAugAST(
   }
 
   map(node);
-}
-
-function isSelection(el: Element) {
-  return el.classList.contains("dcg-mq-selection");
-}
-
-export class MqNodeViaDom {
-  mq: MathQuillField;
-  domNode: Element;
-
-  constructor(mq: MathQuillField, domNode: Element) {
-    this.mq = mq;
-    this.domNode = domNode;
-  }
-
-  parent() {
-    if (this.domNode.classList.contains("dcg-mq-root-block")) {
-      return undefined;
-    }
-    let parent = this.domNode.parentElement;
-    if (!parent) return undefined;
-    if (isSelection(parent)) {
-      parent = parent.parentElement;
-      if (!parent) return undefined;
-    }
-    return new MqNodeViaDom(this.mq, parent);
-  }
-
-  nextSibling(): MqNodeViaDom | undefined {
-    let sibling = this.domNode.nextElementSibling;
-    while (sibling?.classList.contains("dcg-mq-cursor")) {
-      sibling = sibling.nextElementSibling;
-    }
-    const parent = this.domNode.parentElement;
-    if (!sibling && parent && isSelection(parent)) {
-      sibling = parent.nextElementSibling;
-    }
-    if (sibling) {
-      return new MqNodeViaDom(this.mq, sibling);
-    }
-    return undefined;
-  }
-
-  prevSibling(): MqNodeViaDom | undefined {
-    let sibling = this.domNode.previousElementSibling;
-    while (sibling?.classList.contains("dcg-mq-cursor")) {
-      sibling = sibling.previousElementSibling;
-    }
-    const parent = this.domNode.parentElement;
-    if (!sibling && parent && isSelection(parent)) {
-      sibling = parent.previousElementSibling;
-    }
-    if (sibling) {
-      return new MqNodeViaDom(this.mq, sibling);
-    }
-    return undefined;
-  }
-
-  latex() {
-    const span = this.mq.domNodeToSpan(this.domNode);
-    if (!span) return "";
-    return span.latex.slice(span.startIndex, span.endIndex);
-  }
-}
-
-type DomCursor =
-  | { type: "cursor"; el: Element }
-  | { type: "selection"; el: Element; headSide: 1 | -1 };
-
-class MqCursorViaDom {
-  mq: MathQuillField;
-  /**
-   * The domCursor represents the head of the selection.
-   * Since we only have access to DOM nodes, this is a bit awkward.
-   */
-  domCursor: DomCursor;
-
-  constructor(mq: MathQuillField, domCursor: DomCursor) {
-    this.mq = mq;
-    this.domCursor = domCursor;
-  }
-
-  nodeBefore(): MqNodeViaDom | undefined {
-    const sibling =
-      this.domCursor.type === "selection" && this.domCursor.headSide === 1
-        ? this.domCursor.el.lastElementChild
-        : this.domCursor.el.previousElementSibling;
-    if (sibling) return new MqNodeViaDom(this.mq, sibling);
-    else return undefined;
-  }
-
-  nodeAfter(): MqNodeViaDom | undefined {
-    const sibling =
-      this.domCursor.type === "selection" && this.domCursor.headSide === -1
-        ? this.domCursor.el.firstElementChild
-        : this.domCursor.el.nextElementSibling;
-    if (sibling) return new MqNodeViaDom(this.mq, sibling);
-    else return undefined;
-  }
-
-  nodeInDirection(dir: 1 | -1): MqNodeViaDom | undefined {
-    return dir === 1 ? this.nodeAfter() : this.nodeBefore();
-  }
-
-  parentGroup() {
-    const parent = this.domCursor.el.parentElement;
-    if (!parent) throw new Error("Invalid cursor.");
-    return new MqNodeViaDom(this.mq, parent);
-  }
-}
-
-/**
- * Get the head of the selection.
- * When the selection is a point (i.e. just the cursor), it represents that point.
- * Otherwise, this is the head of the selection, which is what moves when you shift-arrow.
- */
-export function getCursorHead(mq: MathQuillField): MqCursorViaDom | undefined {
-  const cursorDom = mq.el().querySelector(".dcg-mq-cursor");
-  if (cursorDom) {
-    return new MqCursorViaDom(mq, { type: "cursor", el: cursorDom });
-  }
-  const selectionDom = mq.el().querySelector(".dcg-mq-selection");
-  if (selectionDom) {
-    const direction = selectionDirection(mq);
-    return new MqCursorViaDom(mq, {
-      type: "selection",
-      el: selectionDom,
-      headSide: direction,
-    });
-  }
-  return undefined;
 }
 
 export function getController(mq: MathQuillField) {
